+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/types/chat/abstracts/extended/ExtendedChannelChat.kt
@@ -1,5 +1,8 @@
 package dev.inmo.tgbotapi.types.chat.abstracts.extended
 
+import dev.inmo.tgbotapi.types.ChatId
 import dev.inmo.tgbotapi.types.chat.abstracts.ChannelChat
 
-interface ExtendedChannelChat : ChannelChat, ExtendedPublicChat
+interface ExtendedChannelChat : ChannelChat, ExtendedPublicChat {
+    val linkedGroupChatId: ChatId?
+}